// gmp-platform/frontend/vendor/online-form-designer/src/projects/app-designer/src/views/online-form/types/designer-type.ts
export enum DesignerType {
  /**
   * 流程表单/流程设计
   */
  ONLINE_FORM = 'onlineForm',
  /**
   * 审核流程模板（设计）
   */
  BIZ_PROCESS_TEMPLATE = 'bizProcessTemplate',
  /**
   * 表单流程模板
   */
  FORM_PROCESS_TEMPLATE = 'formProcessTemplate',
}

export enum PanelType {
  PROP = 'prop',
  PERM = 'perm',
  EVENT = 'event',
}

export enum DesignerMode {
  DESIGN = 'design',
  PREVIEW = 'preview',
}